Used a baseline Gravity Forms Add-on Framework
Started off with the same framework I used to start the vinaudit plugin.
I want to create this in the same time frame if possible. which was
approximately 2 weeks to functionality ready for integrating into other
services.

Since this will not really be integrating into other services I think a
3 week turn around time on this is not out of the question. Styling
things and making them pretty may be another thing all together though.